added unit tests for the functionality of creatinng cheeps.
The following tests are these: Testing that CreateCheep adds cheep successfully. Testing that CreateCheep actually creates new author if needed. Testing that timestamp is made correctly. Testing that cheeps appear in GetCheeps, and finally testing that cheeps appear in author timeline. All of the tests (11) are passing, so we gucci ;D